+++ src/backend/main.ts
@@ -1474,8 +1474,8 @@ ipcMain.handle('syncGOGSaves', async (event, gogSaves, appName, arg) =>
   getGame(appName, 'gog').syncSaves(arg, '', gogSaves)
 )
 
-ipcMain.handle('readGOGGameInfoFile', async (event, appName: string) =>
-  GOGLibrary.get().readInfoFile(appName)
+ipcMain.handle('getGOGLaunchOptions', async (event, appName: string) =>
+  GOGLibrary.get().getLaunchOptions(appName)
 )
 
 ipcMain.handle(
