+++ src/common/typedefs/ipcBridge.d.ts
@@ -28,14 +28,11 @@ import {
   DMQueueElement,
   ConnectivityStatus,
   GamepadActionArgs,
-  ExtraInfo
+  ExtraInfo,
+  LaunchOption
 } from 'common/types'
 import { LegendaryInstallInfo } from 'common/types/legendary'
-import {
-  GOGCloudSavesLocation,
-  GOGGameDotInfoFile,
-  GogInstallInfo
-} from 'common/types/gog'
+import { GOGCloudSavesLocation, GogInstallInfo } from 'common/types/gog'
 
 /**
  * Some notes here:
@@ -241,9 +238,7 @@ interface AsyncIPCFunctions {
   }) => Promise<boolean>
   toggleDXVK: (args: ToolArgs) => Promise<boolean>
   pathExists: (path: string) => Promise<boolean>
-  readGOGGameInfoFile: (
-    appName: string
-  ) => Promise<GOGGameDotInfoFile | undefined>
+  getGOGLaunchOptions: (appName: string) => Promise<LaunchOption[]>
 }
 
 // This is quite ugly & throws a lot of errors in a regular .ts file
